+++ exporter/opentelemetry-exporter-jaeger-proto-grpc/tests/test_jaeger_exporter_protobuf.py
@@ -19,11 +19,11 @@ from unittest import mock
 
 # pylint:disable=no-name-in-module
 # pylint:disable=import-error
-import opentelemetry.exporter.jaeger.proto.gen.model_pb2 as model_pb2
-import opentelemetry.exporter.jaeger.proto.translate as pb_translator
+import opentelemetry.exporter.jaeger.proto.grpc.gen.model_pb2 as model_pb2
+import opentelemetry.exporter.jaeger.proto.grpc.translate as pb_translator
 from opentelemetry import trace as trace_api
-from opentelemetry.exporter.jaeger.proto import JaegerExporter
-from opentelemetry.exporter.jaeger.proto.translate import (
+from opentelemetry.exporter.jaeger.proto.grpc import JaegerExporter
+from opentelemetry.exporter.jaeger.proto.grpc.translate import (
     NAME_KEY,
     VERSION_KEY,
     Translate,